+++ src/main.js
@@ -6,6 +6,7 @@ import SiteMenuView from './view/site-menu-view.js';
 import SortView from './view/sort-view.js';
 import {render, RenderPosition, replace} from './render.js';
 import {generatePoint} from './mock/task.js';
+import PointPresenter from './presenter/point-presenter.js';
 
 const LENGTH_POINTS_ARRAY =15;
 
@@ -25,39 +26,42 @@ const siteFiltersElement = siteBodyElement.querySelector('.trip-controls__filter
 const siteEventsElement = siteBodyElement.querySelector('.trip-events');
 const siteEventsListElement = siteEventsElement .querySelector('.trip-events__list');
 
-const renderTask = (taskListElement, point) => {
-  const taskComponent = new PointListView(point);
-  //console.log('taskComponent',taskComponent);
-  const taskEditComponent = new FormEditView(point);
+// const renderTask = (taskListElement, point) => {
+//   const taskComponent = new PointListView(point);
+//   //console.log('taskComponent',taskComponent);
+//   const taskEditComponent = new FormEditView(point);
 
-  const replaceCardToForm = () => {
-    replace(taskEditComponent, taskComponent);
-  };
+//   const replaceCardToForm = () => {
+//     replace(taskEditComponent, taskComponent);
+//   };
 
-  const replaceFormToCard = () => {
-    replace(taskComponent, taskEditComponent);
-  };
+//   const replaceFormToCard = () => {
+//     replace(taskComponent, taskEditComponent);
+//   };
 
-  const onEscKeyDown = (evt) => {
-    if (evt.key === 'Escape' || evt.key === 'Esc') {
-      evt.preventDefault();
-      replaceFormToCard();
-      document.removeEventListener('keydown', onEscKeyDown);
-    }
-  };
+//   const onEscKeyDown = (evt) => {
+//     if (evt.key === 'Escape' || evt.key === 'Esc') {
+//       evt.preventDefault();
+//       replaceFormToCard();
+//       document.removeEventListener('keydown', onEscKeyDown);
+//     }
+//   };
 
-  taskComponent.setEditClickHandler(() => {
-    replaceCardToForm();
-    document.addEventListener('keydown', onEscKeyDown);
-  });
+//   taskComponent.setEditClickHandler(() => {
+//     replaceCardToForm();
+//     document.addEventListener('keydown', onEscKeyDown);
+//   });
 
-  taskEditComponent.setFormSubmitHandler(() => {
-    replaceFormToCard();
-    document.removeEventListener('keydown', onEscKeyDown);
-  });
+//   taskEditComponent.setFormSubmitHandler(() => {
+//     replaceFormToCard();
+//     document.removeEventListener('keydown', onEscKeyDown);
+//   });
 
-  render(taskListElement, taskComponent, RenderPosition.AFTERBEGIN);
-};
+//   render(taskListElement, taskComponent, RenderPosition.AFTERBEGIN);
+// };
+
+const boardPresenter = new PointPresenter(SiteMenuView);
+boardPresenter.init(POINTS);
 
 if (POINTS.length === 0) {
   render(siteEventsElement, new MessageView(), RenderPosition.BEFOREEND);
